Fold AbstractProtocolScheduleBuilder into ProtocolScheduleBuilder
Now that ProtocolScheduleBuilder is the only subclass
